Standardize on "local-fonts" for permission/policy name
Because we initially thought of this API as potentially being split
into two pieces (local font enumeration and font data access) the spec
has different names for the permission/powerful feature
("font-access") vs. the permissions policy-controlled feature
("local-fonts"). These should be unified.

Per discussion, the preference is to use "local-fonts".

(Since we had a discussion out of band leading on this issue, and it's
so small a change, skipping a PR and landing directly.)